+++ cli/src/cli_util.rs
@@ -56,6 +56,7 @@ use jj_lib::backend::CommitId;
 use jj_lib::backend::MergedTreeId;
 use jj_lib::backend::TreeValue;
 use jj_lib::commit::Commit;
+use jj_lib::config::ConfigNamePathBuf;
 use jj_lib::file_util;
 use jj_lib::fileset;
 use jj_lib::fileset::FilesetDiagnostics;
@@ -145,7 +146,6 @@ use crate::complete;
 use crate::config::new_config_path;
 use crate::config::AnnotatedValue;
 use crate::config::CommandNameAndArgs;
-use crate::config::ConfigNamePathBuf;
 use crate::config::ConfigSource;
 use crate::config::LayeredConfigs;
 use crate::diff_util;
