+++ lib/src/lib.rs
@@ -32,6 +32,7 @@ pub mod annotate;
 pub mod backend;
 pub mod commit;
 pub mod commit_builder;
+pub mod config;
 pub mod conflicts;
 pub mod copies;
 pub mod dag_walk;
